+++ src/buckets.ts
@@ -1,0 +1,23 @@
+type Bucket = number;
+
+interface BucketPair {
+	blockBucket: Bucket;
+	timestampBucket: Bucket;
+}
+
+interface Buckets {
+	currentBuckets: BucketPair;
+	nextBuckets: BucketPair;
+}
+
+const BucketSize = {
+    block: 240,
+    timestamp: 3600,
+}
+
+export {
+    Bucket,
+    BucketPair,
+    Buckets,
+    BucketSize,
+}
